requirements.txt
streamlit
openai
python-dotenv
mcp
tushare
akshare